pyproject.toml
[tool.poetry]
name = "gntech"
version = "0.1.0"
description = "A test for a job position at gntech. Rodrigo Farah is the candidate."
authors = ["Rodrigo Farah <digofarah@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
black = "^25.1.0"
coverage = "^7.8.0"
django-stubs = "^5.1.3"
django-stubs-ext = "^5.1.3"
isort = "^6.0.1"
mypy = "^1.15.0"
psycopg2-binary = "^2.9.10"
pytest = "^8.3.5"
pytest-django = "^4.10.0"
ruff = "^0.11.2"
djangorestframework = "^3.16.0"
python-decouple = "^3.8"
requests = "^2.32.3"

[tool.isort]
profile = "black"
line_length = 88

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "Q", "I"]
ignore = ["E203"]
mypy = true
extend-select = ["C4", "SIM", "TCH"]
target-version = "py310"
fix = true

[tool.mypy]
plugins = [
    "mypy_django_plugin.main"
]
python_version = 3.10
ignore_missing_imports = true
strict = true

[mypy.plugins.django-stubs]
django_settings_module = "project.settings"

[build-system]
requires = ["poetry-core>=1.8"]
build-backend = "poetry.core.masonry.api"